+++ lib/ooxml_parser/pptx_parser/pptx_data/presentation/slide/slide/timing/set_time_node/set_time_node.rb
@@ -1,3 +1,5 @@
+require_relative 'behavior/behavior'
+
 module OoxmlParser
   class SetTimeNode
     attr_accessor :behavior, :to